move tests around
