+++ backend/algorithms/exampleUnselectCourse.json
@@ -6,12 +6,12 @@
             "ACCTA2": 1
         },
         "courses": {
-            "COMP1511": [6, 84],
-            "COMP1521": [6, 57],
-            "COMP1531": [6, 66],
-            "COMP2521": [6, 57],
+            "COMP9242": [6, 81],
             "COMP3231": [6, 61],
-            "COMP9242": [6, 81]
+            "COMP2521": [6, 57],
+            "COMP1531": [6, 66],
+            "COMP1521": [6, 57],
+            "COMP1511": [6, 84]
         },
         "year": 3
     },
